feat: use react-native storage (#25)
* feat: add dependencies

- @react-native-async-storage/async-storage
- uuid-random

* feat: implement insecure solution

* fix: load stored secrets

* fix: test secrets manager

* feat: implement secure solution

* fix: limit test scope

This test case was relying on multiple renders in order to fulfill its
expectations. Multiple renders mean multiple async operations
throughout the components lifecycles which makes everything harder to
predict. The solution here was to then replace the many expectations
relying on the navigation back home after the QR code is scanned with a
single expectation asserting the right function were called with
the right arguments.

* fix: rename hook function

* fix: initialize secrets with fonts

* fix: manage deps with `expo add`